Issue 11323 - imCMS RB 4: Implement CKEditor instead of Xinha. - Largest memory leak fixed. There's an interval that checking if the content is changed every 2 sec. Eventually it couldn't keep up the pace so there started multiple instances of the same function until the JS memory had enough.
I added a global boolean in the function that was changed to true when the function started and back to false when it was completed. And then I made the interval skip to initiate a new call to the function if the boolean was true (busy). Then it tried again on the next cycle, 2 sec later.

To patch other sites (4.0.X with CKEditor):
Copy the last JS tag in the body tag to the swe/eng versions of change_text.jsp

<script type="text/javascript">
...~970 lines ...
</script>

</body>

git-svn-id: https://svn.imcode.com/imcode/imcms/branches/RB-4@24728 bd66a97b-2aff-0310-9095-89ca5cabf5a6